srm,pool: Ensure forward compatibility of srmCopy transfers
JGlobus 2 changes the GlobusGSSCredentialImpl class and breaks backwards
compatibility of serialization. We include an instance of this class in
RemoteGsiftpTransferProtocolInfo. Thus upgrading to JGlobus 2 will break
compatibility with previous versions.

Since we plan to backport JGlobus 2 to dCache 2.2, this is not just a
concern for pool compatibility, but a concern for head node
compatibility too.

Solving this problem is not easy, as we cannot easily force
serialization of the new GlobusGSSCredentialImpl class to become
compatible with the old format. It stands to reason that relying on the
long term serialization stability of JGlobus classes is a bad idea.

I propose the following migration strategy. This patch introduces a
private key field and a certificate chain field to
RemoteGsiftpTransferProtocolInfo. The GSSCrendential field is kept, but
marked deprecated. A readObject() method is added to extract the key and
certificate chain if it is missing. This ensures that the updated
RemoteGsiftpTransferProtocolInfo class is serialization compatible with
the existing version - in either direction.

We merge ths patch into 2.2 and make a new release of that branch, let's
call that release x.

In a future release, y for y > x, we remove the credential field and
upgrade to JGlobus 2. A staged upgrade MUST upgrade the SRM to release
y before any other node, and any other node must be at at least version
x.

y is compatible with x, except for the SRM which must be at y. y is not
compatible with x - 1.

This provides a staged upgrade strategy, both for JGlobus 2 in 2.2 and
for moving forward from 2.2 to 2.6. Obviosly we should try to make y
as big as possible to allow for a smooth transition.

A separate patch for 2.5 and 2.6 will drop the credential field right away.
It will only be compatible with 2.2 pools of release x and newer.

Target: 2.2
Request: 2.2
Require-notes: yes
Require-book: no
Signed-off-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/5225/