srm: Restore RESTORED state
A recent patch removed the RESTORED state from the SRM. This however causes
problems upgrading existing systems as they are unable to map value of state
fields in the database to the enum in the code. This patch fixes the regression
by adding the enum value back.

The patch also fixes a cut'n'paste error message in AccessLatency.

Target: trunk
Require-notes: no
Require-book: no