Use FST for message encoding
This patch introduces FST for payload serialization. FST is an API compatible
drop-in replacement for Java Serialization. It is not wire compatible.
For this reason the tunnel now exchanges an encoding version string and
falls back to Java Serialization if the encoding does not match.

FST is considerably faster than Java serialization and it achieves this
speedup without changes to our message classes. It does however not provide
versioning, which means that ANY change to fields to any class send to
other cells MUST be matched with an increase of the encoding version
string (a constant in the CellMessage class). This will force the tunnel
to reencode using Java Serialization when talking to previous versions
of dCache.

We still have to take care of Java serialization compatibility like before.
Otherwise we would break compatibility with older versions.

The price for the speedup between compatible versions is the cost of
reencoding with Java Serialization between incompatible versions. This
cost can be avoided if we rearchitecture when messages are encoded and
decoded, but currently I have opted for the simpler message reencoding.

Some simple benchmarks have shown an approx. 10 fold increase in raw
serialization/deserialization throughput.