nfs: Add access log
Motivation:

An access log can be quite valuable when debugging client access problems or just
monitoring the activity of a door.

Modification:

Adds an accesss log for the NFS door. The log level can be configured to NONE, MINIMAL and
FULL, although FULL is not currently a full log.

Result:

Adds the property nfs.access-log to configure an access log for NFS doors. Currently the
access log is limited to operations that modify the name space.

Target: trunk
Require-notes: yes
Require-book: yes
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>