webdav: Move non-dCache content to separate port
Motivation:

We have always had a problem with exposing non-dCache content such as static files
and the new API through the webdav as the non-dCache content always shadovs the dCache
name space.

Modification:

Add a second port to the webdav door. Non-dCache content is served from this port.

Makes the WebDAV API configurable. It is disabled by default (can be changed as soon
as it respects the authorization and name space settings of the webdav door).

The $static$ paramter of the template is now a URI object, making it possible to
access components of the URI as eg $static.host$ and $static.port$.

Result:

Added a webdav.net.static-port option. The webdav serves non-dCache content such
as image files from this TCP port. webdav.static-content.uri is obsolete.

The webdav door redirects requests to unknown paths from 3880 to 2880.

Requests to the API are no longer logged in the access log.

Target: trunk
Request: 2.16
Require-notes: yes
Require-book: yes

Reviewed at https://rb.dcache.org/r/9342/