cells: Remove rance in routing manager
Motivation:

Routing manager monitors when routes beeing added and removed. It updates
internal data structures accordingly. When adding a route it looks up the
corresponding tunnel for information about whether the remote is a satellite or
core domain. In route removal this code was copied, however there is no
guarantee that the tunnel still exists at this point and thus we may not clean
the information from the internal data structures.

Modification:

We don't need to look up the tunnel information on remove as all we need
to know is the tunnel address.

Result:

Fixed a race that could lead to corrupted state in the routing manager.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.16

Reviewed at https://rb.dcache.org/r/9748/